@@ -11,61 +13,77 @@
     "url": "https://github.com/JedWatson/react-select.git"
   },
   "dependencies": {
-    "classnames": "^2.2.5",
-    "react-input-autosize": "1.1.0",
-    "react-tether": "^0.5.5"
+    "classnames": "^2.2.4",
+    "lodash": "^4.17.4",
+    "prop-types": "^15.5.8",
+    "react-input-autosize": "^2.0.1"
   },
   "devDependencies": {
-    "babel": "^6.23.0",
-    "babel-eslint": "^7.2.1",
-    "chai": "^3.5.0",
-    "coveralls": "^2.12.0",
-    "eslint": "^3.18.0",
-    "eslint-plugin-react": "^6.10.3",
-    "gulp": "^3.9.1",
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.1",
+    "babel-plugin-external-helpers": "^6.22.0",
+    "babel-plugin-istanbul": "^4.1.4",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.8",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "babel-register": "^6.26.0",
+    "chai": "^4.1.2",
+    "coveralls": "^2.11.12",
+    "create-react-class": "^15.5.2",
+    "cross-env": "^5.0.5",
+    "css-loader": "^0.28.7",
+    "eslint": "^4.6.1",
+    "eslint-plugin-react": "^7.3.0",
+    "extract-text-webpack-plugin": "^3.0.0",
+    "html-loader": "^0.5.1",
+    "html-webpack-plugin": "^2.30.1",
     "isomorphic-fetch": "^2.2.1",
     "istanbul": "^0.4.5",
     "jsdom": "^9.12.0",
-    "mocha": "^3.2.0",
-    "react": "^15.4",
-    "react-addons-shallow-compare": "^15.4",
-    "react-addons-test-utils": "^15.4",
-    "react-component-gulp-tasks": "^0.7.7",
-    "react-dom": "^15.4",
-    "react-gravatar": "^2.6.1",
-    "react-highlight-words": "^0.6.0",
-    "react-virtualized": "^9.4.0",
-    "react-virtualized-select": "^2.4.0",
-    "sinon": "^2.1.0",
-    "unexpected": "^10.26.3",
-    "unexpected-dom": "^3.1.1",
-    "unexpected-react": "^3.5.2",
-    "unexpected-sinon": "^10.7.1"
+    "less": "^2.7.2",
+    "less-loader": "^4.0.5",
+    "less-plugin-clean-css": "^1.5.1",
+    "mocha": "^3.0.2",
+    "nps": "^5.7.1",
+    "nps-utils": "^1.3.0",
+    "nyc": "^11.1.0",
+    "react": "^15.5.0",
+    "react-addons-shallow-compare": "^15.5.0",
+    "react-dom": "^15.5.0",
+    "react-gravatar": "^2.4.5",
+    "react-highlight-words": "^0.8.1",
+    "react-test-renderer": "^15.6.1",
+    "react-virtualized": "^9.9.0",
+    "react-virtualized-select": "^3.1.0",
+    "rollup": "^0.49.2",
+    "rollup-plugin-babel": "^3.0.2",
+    "rollup-plugin-commonjs": "^8.2.0",
+    "rollup-plugin-node-resolve": "^3.0.0",
+    "rollup-plugin-uglify": "^2.0.1",
+    "sinon": "^3.2.1",
+    "style-loader": "^0.18.2",
+    "uglify-es": "^3.0.28",
+    "unexpected": "^10.35.0",
+    "unexpected-dom": "^4.0.0",
+    "unexpected-react": "^4.1.0",
+    "unexpected-sinon": "^10.4.0",
+    "webpack": "^3.5.5",
+    "webpack-dev-server": "^2.7.1"
   },
   "peerDependencies": {
-    "react": "^0.14 || ^15.0.0-rc || ^15.0",
-    "react-dom": "^0.14 || ^15.0.0-rc || ^15.0"
-  },
-  "browserify-shim": {
-    "classnames": "global:classNames",
-    "react": "global:React",
-    "react-dom": "global:ReactDOM",
-    "react-input-autosize": "global:AutosizeInput"
+    "react": "^0.14.9 || ^15.3.0 || ^16.0.0-rc || ^16.0",
+    "react-dom": "^0.14.9 || ^15.3.0 || ^16.0.0-rc || ^16.0"
   },
   "scripts": {
-    "build": "gulp clean && NODE_ENV=production gulp build",
-    "bump": "gulp bump",
-    "bump:major": "gulp bump:major",
-    "bump:minor": "gulp bump:minor",
-    "cover": "istanbul cover _mocha -- -u exports --compilers js:babel/register -R spec",
-    "coveralls": "NODE_ENV=test istanbul cover _mocha --report lcovonly -- -u exports --compilers js:babel/register -R spec && cat coverage/lcov.info | coveralls",
-    "examples": "gulp dev:server",
+    "build": "nps build",
+    "cover": "cross-env NODE_ENV=test nyc --reporter=lcov --reporter=text mocha",
+    "coveralls": "cross-env NODE_ENV=test nyc --reporter=lcov --reporter=text mocha && cat coverage/lcov.info | coveralls",
     "lint": "eslint .",
-    "publish:examples": "NODE_ENV=production gulp publish:examples",
-    "release": "NODE_ENV=production gulp release",
-    "start": "gulp dev",
-    "test": "npm run lint && mocha --compilers js:babel/register",
-    "watch": "gulp watch:lib"
+    "publish:examples": "NODE_ENV=production nps publish",
+    "start": "webpack-dev-server --progress",
+    "test": "NODE_ENV=test mocha --compilers js:babel-core/register"
   },
   "keywords": [
     "combobox",
