@@ -1,95 +1,95 @@
-import React, { Component, PropTypes } from 'react';
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import Select from './Select';
 import stripDiacritics from './utils/stripDiacritics';
-
 const propTypes = {
-	autoload:           React.PropTypes.bool.isRequired,       // automatically call the `loadOptions` prop on-mount; defaults to true
-	cache:              React.PropTypes.any,                      // object to use to cache results; set to null/false to disable caching
-	children:           React.PropTypes.func.isRequired,       // Child function responsible for creating the inner Select component; (props: Object): PropTypes.element
-	ignoreAccents:      React.PropTypes.bool,             // strip diacritics when filtering; defaults to true
-	ignoreCase:         React.PropTypes.bool,                // perform case-insensitive filtering; defaults to true
-	loadingPlaceholder: React.PropTypes.oneOfType([  // replaces the placeholder while options are loading
-		React.PropTypes.string,
-		React.PropTypes.node
+	autoload: PropTypes.bool.isRequired,       // automatically call the `loadOptions` prop on-mount; defaults to true
+	cache: PropTypes.any,                      // object to use to cache results; set to null/false to disable caching
+	children: PropTypes.func.isRequired,       // Child function responsible for creating the inner Select component; (props: Object): PropTypes.element
+	ignoreAccents: PropTypes.bool,             // strip diacritics when filtering; defaults to true
+	ignoreCase: PropTypes.bool,                // perform case-insensitive filtering; defaults to true
+	loadOptions: PropTypes.func.isRequired,    // callback to load options asynchronously; (inputValue: string, callback: Function): ?Promise
+	loadingPlaceholder: PropTypes.oneOfType([  // replaces the placeholder while options are loading
+		PropTypes.string,
+		PropTypes.node
 	]),
-	loadOptions:        React.PropTypes.func.isRequired,    // callback to load options asynchronously; (inputValue: string, callback: Function): ?Promise
-	multi:              React.PropTypes.bool,                     // multi-value input
-	options:            PropTypes.array.isRequired,             // array of options
-	placeholder:        React.PropTypes.oneOfType([         // field placeholder, displayed when there's no value (shared with Select)
-		React.PropTypes.string,
-		React.PropTypes.node
+	multi: PropTypes.bool,                     // multi-value input
+	noResultsText: PropTypes.oneOfType([       // field noResultsText, displayed when no options come back from the server
+		PropTypes.string,
+		PropTypes.node
 	]),
-	noResultsText:      React.PropTypes.oneOfType([       // field noResultsText, displayed when no options come back from the server
-		React.PropTypes.string,
-		React.PropTypes.node
+	onChange: PropTypes.func,                  // onChange handler: function (newValue) {}
+	onInputChange: PropTypes.func,             // optional for keeping track of what is being typed
+	options: PropTypes.array.isRequired,       // array of options
+	placeholder: PropTypes.oneOfType([         // field placeholder, displayed when there's no value (shared with Select)
+		PropTypes.string,
+		PropTypes.node
 	]),
-	onChange:           React.PropTypes.func,                  // onChange handler: function (newValue) {}
-	searchPromptText:   React.PropTypes.oneOfType([    // label to prompt for search input
-		React.PropTypes.string,
-		React.PropTypes.node
+	searchPromptText: PropTypes.oneOfType([    // label to prompt for search input
+		PropTypes.string,
+		PropTypes.node
 	]),
-	onInputChange:      React.PropTypes.func,             // optional for keeping track of what is being typed
-	value:              React.PropTypes.any,                      // initial field value
+	value: PropTypes.any,                      // initial field value
 };
 
 const defaultCache = {};
 
 const defaultProps = {
-	autoload:           true,
-	cache:              defaultCache,
-	children:           defaultChildren,
-	ignoreAccents:      true,
-	ignoreCase:         true,
+	autoload: true,
+	cache: defaultCache,
+	children: defaultChildren,
+	ignoreAccents: true,
+	ignoreCase: true,
 	loadingPlaceholder: 'Loading...',
-	options:            [],
-	searchPromptText:   'Type to search',
+	options: [],
+	searchPromptText: 'Type to search',
 };
 
 export default class Async extends Component {
-	constructor(props, context) {
+	constructor (props, context) {
 		super(props, context);
 
 		this._cache = props.cache === defaultCache ? {} : props.cache;
 
 		this.state = {
+			inputValue: '',
 			isLoading: false,
-			options:   props.options,
+			options: props.options,
 		};
 
-		this._onInputChange = this._onInputChange.bind(this);
+		this.onInputChange = this.onInputChange.bind(this);
 	}
 
-	componentDidMount() {
+	componentDidMount () {
 		const { autoload } = this.props;
 
-		if(autoload) {
+		if (autoload) {
 			this.loadOptions('');
 		}
 	}
 
-	componentWillUpdate(nextProps, nextState) {
-		const propertiesToSync = ['options'];
-		propertiesToSync.forEach((prop) => {
-			if(this.props[prop] !== nextProps[prop]) {
-				this.setState({
-					[prop]: nextProps[prop]
-				});
-			}
-		});
+	componentWillReceiveProps(nextProps) {
+		if (nextProps.options !== this.props.options) {
+			this.setState({
+				options: nextProps.options,
+			});
+		}
 	}
 
-	clearOptions() {
-		this.setState({ options: [] });
+	componentWillUnmount () {
+		this._callback = null;
 	}
 
-	loadOptions(inputValue) {
+	loadOptions (inputValue) {
 		const { loadOptions } = this.props;
-		const cache           = this._cache;
+		const cache = this._cache;
 
-		if(
-				cache &&
-				cache.hasOwnProperty(inputValue)
+		if (
+			cache &&
+			Object.prototype.hasOwnProperty.call(cache, inputValue)
 		) {
+			this._callback = null;
+
 			this.setState({
 				options: cache[inputValue]
 			});
@@ -98,14 +98,14 @@ export default class Async extends Component {
 		}
 
 		const callback = (error, data) => {
-			if(callback === this._callback) {
+			const options = data && data.options || [];
+
+			if (cache) {
+				cache[inputValue] = options;
+			}
+
+			if (callback === this._callback) {
 				this._callback = null;
-
-				const options = data && data.options || [];
-
-				if(cache) {
-					cache[inputValue] = options;
-				}
 
 				this.setState({
 					isLoading: false,
@@ -118,100 +118,88 @@ export default class Async extends Component {
 		this._callback = callback;
 
 		const promise = loadOptions(inputValue, callback);
-		if(promise) {
+		if (promise) {
 			promise.then(
-					(data) => callback(null, data),
-					(error) => callback(error)
+				(data) => callback(null, data),
+				(error) => callback(error)
 			);
 		}
 
-		if(
-				this._callback &&
-				!this.state.isLoading
+		if (
+			this._callback &&
+			!this.state.isLoading
 		) {
 			this.setState({
 				isLoading: true
 			});
 		}
-
-		return inputValue;
 	}
 
-	_onInputChange(inputValue) {
+	onInputChange (inputValue) {
 		const { ignoreAccents, ignoreCase, onInputChange } = this.props;
+		let transformedInputValue = inputValue;
 
-		if(ignoreAccents) {
-			inputValue = stripDiacritics(inputValue);
+		if (ignoreAccents) {
+			transformedInputValue = stripDiacritics(transformedInputValue);
 		}
 
-		if(ignoreCase) {
-			inputValue = inputValue.toLowerCase();
+		if (ignoreCase) {
+			transformedInputValue = transformedInputValue.toLowerCase();
 		}
 
-		if(onInputChange) {
-			onInputChange(inputValue);
+		if (onInputChange) {
+			onInputChange(transformedInputValue);
 		}
 
-		return this.loadOptions(inputValue);
-	}
+		this.setState({ inputValue });
+		this.loadOptions(transformedInputValue);
 
-	inputValue() {
-		if(this.select) {
-			return this.select.state.inputValue;
-		}
-		return '';
+		// Return the original input value to avoid modifying the user's view of the input while typing.
+		return inputValue;
 	}
 
 	noResultsText() {
 		const { loadingPlaceholder, noResultsText, searchPromptText } = this.props;
-		const { isLoading }                                           = this.state;
+		const { inputValue, isLoading } = this.state;
 
-		const inputValue = this.inputValue();
-
-		if(isLoading) {
+		if (isLoading) {
 			return loadingPlaceholder;
 		}
-		if(inputValue && noResultsText) {
+		if (inputValue && noResultsText) {
 			return noResultsText;
 		}
 		return searchPromptText;
 	}
 
-	focus() {
+	focus () {
 		this.select.focus();
 	}
 
-	render() {
-		const { children, loadingPlaceholder, placeholder } = this.props;
-		const { isLoading, options }                        = this.state;
+	render () {
+		const { children, loadingPlaceholder, multi, onChange, placeholder, value } = this.props;
+		const { isLoading, options } = this.state;
 
 		const props = {
 			noResultsText: this.noResultsText(),
-			placeholder:   isLoading ? loadingPlaceholder : placeholder,
-			options:       (isLoading && loadingPlaceholder) ? [] : options,
-			ref:           (ref) => (this.select = ref),
-			onChange:      (newValues) => {
-				if(this.props.multi && this.props.value && (newValues.length > this.props.value.length)) {
-					this.clearOptions();
-				}
-				this.props.onChange(newValues);
-			}
+			placeholder: isLoading ? loadingPlaceholder : placeholder,
+			options: (isLoading && loadingPlaceholder) ? [] : options,
+			ref: (ref) => (this.select = ref),
 		};
 
 		return children({
 			...this.props,
 			...props,
 			isLoading,
-			onInputChange: this._onInputChange
+			onInputChange: this.onInputChange
 		});
 	}
-};
+}
 
-Async.propTypes    = propTypes;
+Async.propTypes = propTypes;
 Async.defaultProps = defaultProps;
 
-function defaultChildren(props) {
+function defaultChildren (props) {
 	return (
-			<Select {...props} />
+		<Select {...props} />
 	);
-};
+}
