@@ -1,94 +1,107 @@
 import React from 'react';
 
-const MultiSelectValueList = React.createClass({
+class MultiSelectValueList extends React.Component {
 
-	displayName: 'MultiSelectValueList',
+  constructor(props) {
+    super(props);
 
-	propTypes: {
-		children:  React.PropTypes.node,
-		className: React.PropTypes.string,
-		disabled:  React.PropTypes.bool,               // disabled prop passed to ReactSelect
-		id:        React.PropTypes.string,                   // Unique id for the value - used for aria
-		onClick:   React.PropTypes.func,                // method to handle click on value label
-		onRemove:  React.PropTypes.func,               // method to handle removal of the value
-		value:     React.PropTypes.object.isRequired     // the option object for this value
-	},
+    this.handleMouseDown      = this.handleMouseDown.bind(this);
+    this.onRemove             = this.onRemove.bind(this);
+    this.handleTouchEndRemove = this.handleTouchEndRemove.bind(this);
+    this.handleTouchMove      = this.handleTouchMove.bind(this);
+    this.handleTouchStart     = this.handleTouchStart.bind(this);
+    this.handleTouchMove      = this.handleTouchMove.bind(this);
+  }
 
-	handleMouseDown (event) {
-		if(event.type === 'mousedown' && event.button !== 0) {
-			return;
-		}
-		if(this.props.onClick) {
-			event.stopPropagation();
-			this.props.onClick(this.props.value, event);
-			return;
-		}
-		if(this.props.value.href) {
-			event.stopPropagation();
-		}
-	},
+  handleMouseDown(event) {
+    event.preventDefault();
+    event.stopPropagation();
 
-	onRemove (event) {
-		event.preventDefault();
-		event.stopPropagation();
-		this.props.onRemove(this.props.value);
-	},
+    if(event.type === 'mousedown' && event.button !== 0) {
+      return;
+    }
 
-	handleTouchEndRemove (event){
-		// Check if the view is being dragged, In this case
-		// we don't want to fire the click event (because the user only wants to scroll)
-		if(this.dragging) {
-			return;
-		}
+    if(this.props.onClick) {
+      event.stopPropagation();
+      this.props.onClick(this.props.value, event);
+      return;
+    }
 
-		// Fire the mouse events
-		this.onRemove(event);
-	},
+    if(this.props.value.href) {
+      event.stopPropagation();
+    }
+  }
 
-	handleTouchMove (event) {
-		// Set a flag that the view is being dragged
-		this.dragging = true;
-	},
+  onRemove(event) {
+    event.preventDefault();
+    event.stopPropagation();
+    this.props.onRemove(this.props.value);
+  }
 
-	handleTouchStart (event) {
-		// Set a flag that the view is not being dragged
-		this.dragging = false;
-	},
+  handleTouchEndRemove(event) {
+    // Check if the view is being dragged, In this case
+    // we don't want to fire the click event (because the user only wants to scroll)
+    if(this.dragging) {
+      return;
+    }
 
-	renderRemoveIcon () {
-		if(this.props.disabled || !this.props.onRemove) {
-			return;
-		}
-		return (
-				<span
-						style={{ float: 'right' }}
-						className="icon-cross3"
-						aria-hidden="true"
-						onMouseDown={this.onRemove}
-						onTouchEnd={this.handleTouchEndRemove}
-						onTouchStart={this.handleTouchStart}
-						onTouchMove={this.handleTouchMove} >
-			</span>
-		);
-	},
+    // Fire the mouse events
+    this.onRemove(event);
+  }
 
-	renderLabel () {
-		return (
-				<span style={{ width: '95%' }} >{ this.props.children }</span>
-		);
-	},
+  handleTouchMove(event) {
+    // Set a flag that the view is being dragged
+    this.dragging = true;
+  }
 
-	render () {
-		return (
-				<div
-						className={'Select--multiItem ' + this.props.className}
-						title={this.props.value.title} >
-					{this.renderLabel()}
-					{this.renderRemoveIcon()}
-				</div>
-		);
-	}
+  handleTouchStart(event) {
+    // Set a flag that the view is not being dragged
+    this.dragging = false;
+  }
 
-});
+  renderRemoveIcon() {
+    if(this.props.disabled || !this.props.onRemove) {
+      return;
+    }
+    return (
+      <span
+        style={{ float: 'right' }}
+        className="icon-cross3"
+        aria-hidden="true"
+        onMouseDown={this.onRemove}
+        onTouchEnd={this.handleTouchEndRemove}
+        onTouchStart={this.handleTouchStart}
+        onTouchMove={this.handleTouchMove} >
+			</span >
+    );
+  }
 
-module.exports = MultiSelectValueList;
+  renderLabel() {
+    return (
+      <span style={{ width: '95%' }} >{this.props.children}</span >
+    );
+  }
+
+  render() {
+    return (
+      <div
+        className={'Select--multiItem ' + this.props.className}
+        title={this.props.value.title} >
+        {this.renderLabel()}
+        {this.renderRemoveIcon()}
+      </div >
+    );
+  }
+}
+
+MultiSelectValueList.propTypes = {
+  children:  React.PropTypes.node,
+  className: React.PropTypes.string,
+  disabled:  React.PropTypes.bool,               // disabled prop passed to ReactSelect
+  id:        React.PropTypes.string,                   // Unique id for the value - used for aria
+  onClick:   React.PropTypes.func,                // method to handle click on value label
+  onRemove:  React.PropTypes.func,               // method to handle removal of the value
+  value:     React.PropTypes.object.isRequired     // the option object for this value
+};
+
+export default MultiSelectValueList;
