@@ -48,4 +48,4 @@ function menuRenderer ({
 	});
 }
 
-module.exports = menuRenderer;
+export default menuRenderer;
