@@ -1,24 +1,29 @@
-"use strict";
+'use strict';
 
 Object.defineProperty(exports, "__esModule", {
 	value: true
 });
-exports["default"] = arrowRenderer;
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
+exports.default = arrowRenderer;
 
 var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
 
+var _propTypes = require('prop-types');
+
+var _propTypes2 = _interopRequireDefault(_propTypes);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
 function arrowRenderer(_ref) {
 	var onMouseDown = _ref.onMouseDown;
 
-	return _react2["default"].createElement("span", {
-		className: "Select-arrow",
+	return _react2.default.createElement('span', {
+		className: 'Select-arrow',
 		onMouseDown: onMouseDown
 	});
-}
+};
 
-;
-module.exports = exports["default"];
+arrowRenderer.propTypes = {
+	onMouseDown: _propTypes2.default.func
+};
