@@ -1,16 +1,20 @@
 'use strict';
 
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
+Object.defineProperty(exports, "__esModule", {
+	value: true
+});
 
 var _stripDiacritics = require('./stripDiacritics');
 
 var _stripDiacritics2 = _interopRequireDefault(_stripDiacritics);
 
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
 function filterOptions(options, filterValue, excludeOptions, props) {
 	var _this = this;
 
 	if (props.ignoreAccents) {
-		filterValue = (0, _stripDiacritics2['default'])(filterValue);
+		filterValue = (0, _stripDiacritics2.default)(filterValue);
 	}
 
 	if (props.ignoreCase) {
@@ -28,8 +32,8 @@ function filterOptions(options, filterValue, excludeOptions, props) {
 		var valueTest = String(option[props.valueKey]);
 		var labelTest = String(option[props.labelKey]);
 		if (props.ignoreAccents) {
-			if (props.matchProp !== 'label') valueTest = (0, _stripDiacritics2['default'])(valueTest);
-			if (props.matchProp !== 'value') labelTest = (0, _stripDiacritics2['default'])(labelTest);
+			if (props.matchProp !== 'label') valueTest = (0, _stripDiacritics2.default)(valueTest);
+			if (props.matchProp !== 'value') labelTest = (0, _stripDiacritics2.default)(labelTest);
 		}
 		if (props.ignoreCase) {
 			if (props.matchProp !== 'label') valueTest = valueTest.toLowerCase();
@@ -39,4 +43,4 @@ function filterOptions(options, filterValue, excludeOptions, props) {
 	});
 }
 
-module.exports = filterOptions;
+exports.default = filterOptions;
