@@ -1,6 +1,8 @@
 'use strict';
 
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
+Object.defineProperty(exports, "__esModule", {
+	value: true
+});
 
 var _classnames = require('classnames');
 
@@ -10,33 +12,35 @@ var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
 
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
 function menuRenderer(_ref) {
-	var focusedOption = _ref.focusedOption;
-	var instancePrefix = _ref.instancePrefix;
-	var labelKey = _ref.labelKey;
-	var onFocus = _ref.onFocus;
-	var onSelect = _ref.onSelect;
-	var optionClassName = _ref.optionClassName;
-	var optionComponent = _ref.optionComponent;
-	var optionRenderer = _ref.optionRenderer;
-	var options = _ref.options;
-	var valueArray = _ref.valueArray;
-	var valueKey = _ref.valueKey;
-	var onOptionRef = _ref.onOptionRef;
+	var focusedOption = _ref.focusedOption,
+	    instancePrefix = _ref.instancePrefix,
+	    labelKey = _ref.labelKey,
+	    onFocus = _ref.onFocus,
+	    onSelect = _ref.onSelect,
+	    optionClassName = _ref.optionClassName,
+	    optionComponent = _ref.optionComponent,
+	    optionRenderer = _ref.optionRenderer,
+	    options = _ref.options,
+	    valueArray = _ref.valueArray,
+	    valueKey = _ref.valueKey,
+	    onOptionRef = _ref.onOptionRef;
 
 	var Option = optionComponent;
 
 	return options.map(function (option, i) {
 		var isSelected = valueArray && valueArray.indexOf(option) > -1;
 		var isFocused = option === focusedOption;
-		var optionClass = (0, _classnames2['default'])(optionClassName, {
+		var optionClass = (0, _classnames2.default)(optionClassName, {
 			'Select-option': true,
 			'is-selected': isSelected,
 			'is-focused': isFocused,
 			'is-disabled': option.disabled
 		});
 
-		return _react2['default'].createElement(
+		return _react2.default.createElement(
 			Option,
 			{
 				className: optionClass,
@@ -49,12 +53,12 @@ function menuRenderer(_ref) {
 				onSelect: onSelect,
 				option: option,
 				optionIndex: i,
-				ref: function (ref) {
-					onOptionRef(ref, isFocused);
+				ref: function ref(_ref2) {
+					onOptionRef(_ref2, isFocused);
 				} },
 			optionRenderer(option, i)
 		);
 	});
 }
 
-module.exports = menuRenderer;
+exports.default = menuRenderer;
