@@ -17,10 +17,10 @@ var expect = unexpected
 
 var React = require('react');
 var ReactDOM = require('react-dom');
-var TestUtils = require('react-addons-test-utils');
+var TestUtils = require('react-dom/test-utils');
 var sinon = require('sinon');
 
-var Select = require('../src/Select');
+var Select = require('../src');
 
 describe('Async', () => {
 	let asyncInstance, asyncNode, filterInputNode, loadOptions;
@@ -31,8 +31,8 @@ describe('Async', () => {
 			<Select.Async
 				autoload={false}
 				openOnFocus
-				{...props}
 				loadOptions={loadOptions}
+				{...props}
 			/>
 		);
 		asyncNode = ReactDOM.findDOMNode(asyncInstance);
@@ -144,6 +144,35 @@ describe('Async', () => {
 			});
 			typeSearchText('te');
 			return expect(asyncNode.textContent, 'to contain', 'Loading');
+		});
+		
+		it('caches the result of all option fetches', (cb) => {
+			const res = {
+				t: createOptionsResponse(['t']),
+				te: createOptionsResponse(['te']),
+				tes: createOptionsResponse(['tes']),
+			};
+			function loadOptions (input, resolve) {
+				const delay = 10 * (3 - input.length);
+				setTimeout(function() {
+					resolve(null, res[input]);
+				}, delay);
+			}
+			createControl({
+				loadOptions,
+			});
+			const instance = asyncInstance;
+			typeSearchText('t');
+			typeSearchText('te');
+			typeSearchText('tes');
+
+			// TODO: How to test this?
+			setTimeout(function() {
+				expect(instance._cache.t, 'to equal', res.t.options);
+				expect(instance._cache.te, 'to equal', res.te.options);
+				expect(instance._cache.tes, 'to equal', res.tes.options);
+				cb();
+			}, 30);
 		});
 
 		describe('with callbacks', () => {
@@ -319,6 +348,15 @@ describe('Async', () => {
 			typeSearchText('WÄRE');
 			expect(loadOptions, 'was called with', 'WÄRE');
 		});
+
+		it('does not mutate the user input', () => {
+			createControl({
+				ignoreAccents: false,
+				ignoreCase: true
+			});
+			typeSearchText('A');
+			expect(asyncNode.textContent.substr(asyncNode.textContent.indexOf('}') + 1), 'to begin with', 'A');
+		});
 	});
 
 	describe('with ignore case and ignore accents', () => {
@@ -377,26 +415,30 @@ describe('Async', () => {
 			});
 
 			describe('if noResultsText has been provided', () => {
-				it('returns the noResultsText', () => {
-					asyncInstance.select = { state: { inputValue: 'asdf' } };
-					expect(asyncInstance.noResultsText(), 'to equal', 'noResultsText');
+				it('returns the noResultsText', (cb) => {
+					asyncInstance.setState({
+						inputValue: 'asfd',
+					}, () => {
+						expect(asyncInstance.noResultsText(), 'to equal', 'noResultsText');
+						cb();
+					});
 				});
 			});
 
 			describe('if noResultsText is empty', () => {
-				beforeEach((cb) => {
+				beforeEach(() => {
 					createControl({
 						searchPromptText: 'searchPromptText',
 						loadingPlaceholder: 'loadingPlaceholder'
 					});
-					asyncInstance.setState({
-						isLoading: false,
-						inputValue: 'asdfkljhadsf'
-					}, cb);
 				});
-				it('falls back to searchPromptText', () => {
-					asyncInstance.select = { state: { inputValue: 'asdf' } };
-					expect(asyncInstance.noResultsText(), 'to equal', 'searchPromptText');
+				it('falls back to searchPromptText', (cb) => {
+					asyncInstance.setState({
+						inputValue: 'asfd',
+					}, () => {
+						expect(asyncInstance.noResultsText(), 'to equal', 'searchPromptText');
+						cb();
+					});
 				});
 			});
 		});
@@ -438,13 +480,29 @@ describe('Async', () => {
 	describe('.focus()', () => {
 		beforeEach(() => {
 			createControl({});
+			TestUtils.Simulate.blur(filterInputNode);
 		});
 
 		it('focuses the search input', () => {
-			var input = asyncNode.querySelector('input');
-			expect(input, 'not to equal', document.activeElement);
+			expect(filterInputNode, 'not to equal', document.activeElement);
 			asyncInstance.focus();
-			expect(input, 'to equal', document.activeElement);
+			expect(filterInputNode, 'to equal', document.activeElement);
+		});
+	});
+
+
+	describe('props sync test', () => {
+		it('should update options on componentWillReceiveProps', () => {
+			createControl({
+			});
+			asyncInstance.componentWillReceiveProps({
+				options: [{
+					label: 'bar',
+					value: 'foo',
+				}]
+			});
+			expect(asyncNode.querySelectorAll('[role=option]').length, 'to equal', 1);
+			expect(asyncNode.querySelector('[role=option]').textContent, 'to equal', 'bar');
 		});
 	});
 });
