@@ -20,10 +20,9 @@ jsdomHelper();
 
 var React = require('react');
 var ReactDOM = require('react-dom');
-var TestUtils = require('react-addons-test-utils');
+var TestUtils = require('react-dom/test-utils');
 
-var Select = require('../src/Select');
-var Value = require('../src/Value');
+var Select = require('../src').default;
 
 // The displayed text of the currently selected item, when items collapsed
 var DISPLAYED_SELECTION_SELECTOR = '.Select-value';
@@ -40,7 +39,6 @@ class PropsWrapper extends React.Component {
 		super(props);
 		this.state = props || {};
 	}
-
 	setPropsForChild(props) {
 		this.setState(props);
 	}
@@ -691,6 +689,11 @@ describe('Select', () => {
 			expect(onChange, 'was called with', 0);
 		});
 
+        it('displays the X button for 0 value', () => {
+            wrapper.setPropsForChild({ value: 0 });
+            expect(ReactDOM.findDOMNode(instance).querySelector('.Select-clear'), 'not to equal', undefined);
+        });
+
 		describe('with multi=true', () => {
 
 			beforeEach(() => {
@@ -961,6 +964,11 @@ describe('Select', () => {
 			expect(ReactDOM.findDOMNode(instance), 'queried for first', DISPLAYED_SELECTION_SELECTOR,
 					'to have text', 'No');
 		});
+
+        it('displays the X button for false value', () => {
+            wrapper.setPropsForChild({ value: false });
+            expect(ReactDOM.findDOMNode(instance).querySelector('.Select-clear'), 'not to equal', undefined);
+        });
 
 		describe('with multi=true', () => {
 
@@ -1840,6 +1848,19 @@ describe('Select', () => {
 			]);
 		});
 
+		it('removes the last item with backspace', () => {
+
+			wrapper.setPropsForChild({
+				multi: false,
+				value: 'one'
+			});
+			onChange.reset();  // Ignore previous onChange calls
+
+			pressBackspace();
+
+			expect(onChange, 'was called with', null);
+		});
+
 		it('doesn\'t show the X if clearableValue=false', () => {
 
 			setValueProp(['two']);
@@ -1936,7 +1957,8 @@ describe('Select', () => {
 				value: '',
 				options: options,
 				searchable: false,
-				multi: true
+				multi: true,
+				closeOnSelect: false,
 			}, {
 				wireUpOnChangeToValue: true
 			});
@@ -1956,8 +1978,8 @@ describe('Select', () => {
 			clickArrowToOpen();
 
 			expect(instance,
-				'with event mouseDown', 'on', <div className="Select-option">Two</div>,
-				'with event mouseDown', 'on', <div className="Select-option">One</div>,
+				'with event', 'mouseDown', 'on', <div className="Select-option">Two</div>,
+				'with event', 'mouseDown', 'on', <div className="Select-option">One</div>,
 				'to contain',
 				<span className="Select-multi-value-wrapper">
 					<div><span className="Select-value-label">Two</span></div>
@@ -2456,7 +2478,7 @@ describe('Select', () => {
 				expect(spyFilterOption, 'was called with', defaultOptions[1], '');
 				expect(spyFilterOption, 'was called with', defaultOptions[2], '');
 				expect(spyFilterOption, 'was called with', defaultOptions[3], '');
-			});
+			}).timeout(5000);
 
 			describe('when entering text', () => {
 
@@ -2474,7 +2496,7 @@ describe('Select', () => {
 					expect(spyFilterOption, 'was called with', defaultOptions[1], 'xyz');
 					expect(spyFilterOption, 'was called with', defaultOptions[2], 'xyz');
 					expect(spyFilterOption, 'was called with', defaultOptions[3], 'xyz');
-				});
+				}).timeout(5000);
 
 				it('only shows the filtered option', () => {
 
@@ -2939,7 +2961,7 @@ describe('Select', () => {
 
 				instance = createControl({
 					options: defaultOptions,
-					onFocus: onFocus
+					onFocus: onFocus,
 				});
 			});
 
@@ -2949,27 +2971,51 @@ describe('Select', () => {
 			});
 		});
 
-		describe('openAfterFocus', () => {
-
-			var openAfterFocus;
-
-			beforeEach(() => {
-				openAfterFocus = sinon.spy();
-
+		describe('openOnClick', () => {
+			it('should open the menu on click when true', () => {
 				instance = createControl({
 					options: defaultOptions,
-					openAfterFocus: true
+					openOnClick: true,
+				}, {
+					initialFocus: false,
 				});
+				TestUtils.Simulate.mouseDown(ReactDOM.findDOMNode(instance).querySelector('.Select-control'), { button: 0 });
+				// need to simulate the focus on the input, it does not happen in jsdom
+				findAndFocusInputControl();
+				expect(instance.state.isOpen, 'to be true');
 			});
 
-			it('should show the options when focused', () => {
-				instance.focus();
+			it('should not open the menu on click when false', () => {
+				instance = createControl({
+					options: defaultOptions,
+					openOnClick: false,
+				}, {
+					initialFocus: false,
+				});
+				TestUtils.Simulate.mouseDown(ReactDOM.findDOMNode(instance).querySelector('.Select-control'), { button: 0 });
+				// need to simulate the focus on the input, it does not happen in jsdom
+				findAndFocusInputControl();
+				expect(instance.state.isOpen, 'to be falsy');
+			});
+		});
 
-				if (instance.state.isFocused && instance.state.openAfterFocus) {
-					expect(instance.state.isOpen, 'to equal', true);
-				}
+		describe('openOnFocus', () => {
+			// Note: createControl automatically focuses the control
+			it('should open the menu on focus when true', () => {
+				instance = createControl({
+					options: defaultOptions,
+					openOnFocus: true,
+				});
+				expect(instance.state.isOpen, 'to be true');
 			});
 
+			it('should not open the menu on focus when false', () => {
+				instance = createControl({
+					options: defaultOptions,
+					openOnFocus: false,
+				});
+				expect(instance.state.isOpen, 'to be falsy');
+			});
 		});
 
 		describe('onValueClick', () => {
@@ -3775,7 +3821,7 @@ describe('Select', () => {
 
 			it('sets the haspopup and expanded to true when menu is shown', () => {
 				expect(instance,
-					'with event keyDown', ARROW_DOWN, 'on', <div className="Select-control" />,
+					'with event', 'keyDown', ARROW_DOWN, 'on', <div className="Select-control" />,
 					'to contain', <input role="combobox" aria-haspopup="true" aria-expanded="true" />);
 			});
 
@@ -3803,6 +3849,19 @@ describe('Select', () => {
 					'to contain', <input role="combobox" aria-labelledby="test-label-id" />);
 			});
 
+			it('passes through the aria-describedby prop', () => {
+
+				instance = createControl({
+					options: defaultOptions,
+					value: 'one',
+					'aria-describedby': 'test-label1-id test-label2-id'
+				});
+
+				expect(instance,
+					'to contain', <input role="combobox" aria-describedby="test-label1-id test-label2-id" />);
+			});
+
+
 			it('passes through the aria-label prop', () => {
 
 				instance = createControl({
@@ -3829,7 +3888,8 @@ describe('Select', () => {
 						{ value: 'five', label: 'label five' }
 					],
 					value: [ 'three', 'two' ],
-					multi: true
+					multi: true,
+					closeOnSelect: false,
 				}, {
 					wireUpOnChangeToValue: true
 				});
@@ -3845,7 +3905,7 @@ describe('Select', () => {
 
 			it('hides the `press backspace to remove` message on blur', () => {
 				expect(instance,
-					'with event blur', 'on', <input role="combobox" />,
+					'with event', 'blur', 'on', <input role="combobox" />,
 					'not to contain',
 					<span className="Select-aria-only" aria-live="assertive">
 						Press backspace to remove label two
@@ -3865,8 +3925,8 @@ describe('Select', () => {
 			it('updates the active descendant after a selection', () => {
 
 				return expect(wrapper,
-					'with event keyDown', ARROW_DOWN, 'on', <div className="Select-control" />,
-					'with event keyDown', KEY_ENTER, 'on', <div className="Select-control" />,
+					'with event', 'keyDown', ARROW_DOWN, 'on', <div className="Select-control" />,
+					'with event', 'keyDown', KEY_ENTER, 'on', <div className="Select-control" />,
 					'queried for', <input role="combobox" />)
 					.then(input => {
 
@@ -3891,6 +3951,45 @@ describe('Select', () => {
 			expect(input, 'not to equal', document.activeElement);
 			instance.focus();
 			expect(input, 'to equal', document.activeElement);
+		});
+	});
+
+	describe('with autoFocus', () => {
+		it('focuses select input on mount', () => {
+			wrapper = createControl({
+				autoFocus: true,
+				options: defaultOptions,
+			});
+			var input = ReactDOM.findDOMNode(instance.input).querySelector('input');
+			expect(input, 'to equal', document.activeElement);
+		});
+		it('with autofocus as well, calls focus() only once', () => {
+			wrapper = createControl({
+				autofocus: true,
+				autoFocus: true,
+				options: defaultOptions,
+			});
+			var focus = sinon.spy(instance, 'focus');
+			instance.componentDidMount();
+			expect(focus, 'was called once');
+		});
+	});
+	describe('with autofocus', () => {
+		it('focuses the select input on mount', () => {
+			wrapper = createControl({
+				autofocus: true,
+				options: defaultOptions,
+			});
+			var input = ReactDOM.findDOMNode(instance.input).querySelector('input');
+			expect(input, 'to equal', document.activeElement);
+		});
+		it('calls console.warn', () => {
+			var warn = sinon.spy(console, 'warn');
+			wrapper = createControl({
+				autofocus: true,
+				options: defaultOptions,
+			});
+			expect(warn, 'was called once');
 		});
 	});
 });
